Reorganise files into dev folder to highlight that it wont be carried over to AEM
