+++ test/components/info/dev/info.data.js
@@ -1,6 +1,6 @@
-const { retrieveJsonData } = require('../../../tools/loadJsonData'),
+const { retrieveJsonData } = require('../../../../tools/loadJsonData'),
 
-  jsonFilePath = 'test/components/info/info.json',
+  jsonFilePath = 'test/components/info/dev/info.json',
   jsonApiUrl = null;
 
 module.exports = new Promise(async (resolve) => {
