+++ test/components/info/dev/info.js
