+++ test/components/info/dev/info.json
