+++ test/components/info/dev/info.scss
