+++ test/components/info2/dev/info2.data.js
@@ -1,6 +1,6 @@
-const { retrieveJsonData } = require('../../../tools/loadJsonData'),
+const { retrieveJsonData } = require('../../../../tools/loadJsonData'),
 
-  jsonFilePath = 'test/components/info2/info2.json',
+  jsonFilePath = 'test/components/info2/dev/info2.json',
   jsonApiUrl = null;
 
 module.exports = new Promise(async (resolve) => {
