+++ test/components/info2/dev/info2.scss
